refactor: useInput 디렉터리 구조 변경
이슈 번호: #31
