+++ frontend/src/pages/LoginPage/components/LoginForm.tsx
@@ -1,4 +1,4 @@
-import useInput from 'hooks/useInput/useInput';
+import useInput from 'hooks/useInput';
 import React from 'react';
 import { Link } from 'react-router-dom';
